v1.1.4

+++ target.py
@@ -83,14 +83,18 @@ def target(exoplanet, curves = 1, dtype = 'nasa'):
         https://mast.stsci.edu/portal/Mashup/Clients/Mast/Portal.html
     '''
     #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
+    tess_filter = getcwd()+'/data/Filters/TESS_filter.csv'
     # Filter Setup
-    cols = ['filter_idx', 'low_wl', 'high_wl']
-    df = DataFrame(columns = cols)
-    df = df.append([{'filter_idx':0,
-                     'low_wl':getcwd()+'/data/TESS_Filter/TESS_filter.csv'}], 
-                                                          ignore_index = True)
-    df.to_csv(r'data/TESS_filter_path.csv', index = False, header = True)
-    print('Assigned filter paths for '+exoplanet+'..')
+    if not path.exists(tess_filter):
+        cols = ['filter_idx', 'low_wl', 'high_wl']
+        df = DataFrame(columns = cols)
+        df = df.append([{'filter_idx':0,
+                         'low_wl':tess_filter}], ignore_index = True)
+        df.to_csv(r'data/TESS_filter_path.csv', index = False, header = True)
+        print('Assigned TESS filter paths for '+exoplanet+'..')
+    else:
+        print('TESS filter paths already assigned..')
+        pass
     #~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~#
     # Download EU Data
     makedirs('data', exist_ok = True)
